requirements.txt
asgiref==3.5.2
distlib==0.3.5
dj-database-url==1.0.0
Django==4.0.6
django-crispy-forms==1.14.0
django-filter==22.1
django-on-heroku==1.1.2
filelock==3.7.1
gunicorn==20.1.0
mysql-connector-python==8.0.29
mysqlclient==2.1.1
Pillow==9.2.0
platformdirs==2.5.2
protobuf==4.21.4
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.15.1

